requirements.txt
# ---- Dashboard server ----
fastapi>=0.110
uvicorn[standard]>=0.27
watchdog>=4.0
pywinpty>=2.0       ; sys_platform == "win32"

# ---- Voice assistant ("hey grassy") ----
python-dotenv>=1.0
requests>=2.31
sounddevice>=0.5
numpy>=1.26
soundfile>=0.12
pynput>=1.7
faster-whisper>=1.1
webrtcvad-wheels>=2.0
edge-tts>=7.0
pygame>=2.5

# ---- Optional GPU acceleration for Whisper STT ----
# Install these for CUDA — voice-wakeword.py auto-falls-back to CPU if absent.
nvidia-cublas-cu12 ; sys_platform == "win32"
nvidia-cudnn-cu12  ; sys_platform == "win32"

# ---- Touch Grass mode: cam preview + SigLIP 2 detector ----
# For CUDA, install torch from the PyTorch wheel index first:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu124
# Plain `pip install -r requirements.txt` installs the CPU build of torch
# (still works, ~10x slower per frame).
opencv-python>=4.9
torch>=2.4
torchvision>=0.19
transformers>=4.45
pillow>=10.0
accelerate>=0.30
